src/web_app/requirements.txt
requests==2.18.4
python-dateutil==2.6.1
boto3==1.7.6
fastapi==0.42.0
mangum==0.6.17